Improve the links table and related ...
Make the links clickable, open the link in new tab.
Make the copy button copy the shortcut link and make the user now that the copy process was ok.
Remove clicks from the table as the API does not increment the clicks on new clicks.
Prepend https://usict.acm.org/link/ to the code used in the shortened link.